# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.3.6"
  title: BetProtocol MicroService
# during dev, should point to your local machine
host: localhost:8080
# basePath prefixes all resource paths 
basePath: /api
# 

schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /status/post:
    x-swagger-router-controller: ping
    post:
      tags:
      - ping
      summary: Test Ping
      operationId: pingPost
      description: Test Ping for post.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: ping
          description: Ping Object for test
          required: true
          schema:
            $ref: '#/definitions/PingPost'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/credit:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Update Wallet Amount for the user
      operationId: webhookDeposit
      description: Update Wallet Amount for the user
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name : id
          type : string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/requestWithdraw: 
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Request for User Withdraw
      operationId: requestWithdraw
      description: Update Wallet Amount for the User
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User Information
          required: true
          schema:
            $ref: '#/definitions/UserWithdrawRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/transactions:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Get Transactions From User
      operationId: getTransactions
      description: Get Data From User Transactions 
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: Get Data From User Transactions
          required: true
          schema:
            $ref: '#/definitions/GetTransactionsRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/cancelWithdraw:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Request for User Cancel Withdraw
      operationId: cancelWithdraw
      description: Request Cancel Withdraw
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User Information
          required: true
          schema:
            $ref: '#/definitions/UserCancelWithdrawRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /app/wallet/currency/add:
    x-swagger-router-controller: app
    post:
      tags:
      - app
      summary: Add Currency Wallet
      operationId: addCurrencyWallet
      description: Add Currency Wallet
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: app
          description: App Information
          required: true
          schema:
            $ref: '#/definitions/AppAddCurrencyWallet'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /app/address/get : 
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Get Address Deposit for User
      operationId: getDepositAddress
      description: Change Footer
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: app
          description: App Information
          required: true
          schema:
            $ref: '#/definitions/AppUserDepositAddressRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
  
# complex objects have schema definitions
definitions:
  PingPost:
    type: object
    required :
      - type
    properties:
      type:
        type : string
        example: "24234g234g23g"
        minLength: 1
        description: |
          Type Auth
      app:
        type : string
        example: "24234g234g23g"
        description: |
          APP ID
  UserWithdrawRequest : 
    type: object
    required:
      - app
      - user
      - address
      - tokenAmount
      - nonce
      - currency
    properties:
      tokenAmount : 
        type : number
        example: 203
        description: |
          Amount of Tokens to be Withdraw
      nonce:
        type : number
        example: 3245
        description: |
          Use Nonce
      app :
        type : string
        format : name
        example : "5cuedgq7e6t334t34" 
        description: |
          App ID
      user :
        type : string
        format: name
        example : "5cuedgq7e6t334t34" 
        description: |
          User ID
      address :
        type : string
        format: name
        example: "0xe8rghwe8rghwe9r87ynw987eyncwre"
        description: |
          User Address
      currency :
        type : string
        format: name
        example: "0xe8rghwe8rghwe9r87ynw987eyncwre"
        description: |
          User Address
      isAffiliate:
            type : boolean
            example: true
  GetTransactionsRequest : 
    type: object
    required:
      - app
      - user
    properties:
      app :
        type : string
        format : name
        example : "5cuedgq7e6t334t34" 
        description: |
          App ID
      size :
        type : number
      user :
        type : string
        format : name
        example : "5cuedgq7e6t334t34" 
        description: |
          User Id
      offset :
        type : number
  UserCancelWithdrawRequest :
    type: object
    required:
      - app
      - user
      - withdraw_id
      - currency
      - note
    properties:
      app :
        type : string
        format : name
        example : "5cuedgq7e6t334t34"
        description: |
          App ID
      user :
        type : string
        format: name
        example : "5cuedgq7e6t334t34"
        description: |
          User ID
      withdraw_id :
        type : string
        format: name
        example: "65f76st8vdsb3181"
        description: |
            Withdraw Id
      currency :
        type : string
        format: name
        example: "0xe8rghwe8rghwe9r87ynw987eyncwre"
        description: |
          User Address
      note :
        type : string
        format: name
        example: "Test Test"
        description: |
          Note text
  AppAddCurrencyWallet : 
    type: object
    required:
      - app
      - currency_id
      - admin
    properties:
      app:
        type : string
        format: name
        example: "wertwertwertwert"
        description: |
          App id
      currency_id:
        type : string
        format: name
        example: "wertwertwertwert"
        description: |
          Currency id
      admin:
        type : string
        format: name
        example: "wertwertwertwert"
        description: |
          Admin id
  AppUserDepositAddressRequest : 
    type: object
    required:
      - app
      - currency
      - id
    properties:
      app :
        type : string
        format : name
        example : "5cuedgq7e6t334t34" 
        description: |
          App ID
      currency :
        type : string
        format : name
        example : "5cuedgq7e6t334t34" 
        description: |
          Currency ID
      id :
        type : string
        format : name
        example : "5cuedgq7e6t334t34" 
        description: |
          User ID
  GeneralResponse : 
    type: object
    required:
      - data
      - meta            
    properties:
      data:
        type : object
        format: data
        description: |
          Unique identifier of the user, besides the ID
      meta:
        type : object
        format: meta
        description: |
          User Real Name
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string